+++ app/src/main/java/com/makeus/eatoo/src/home/group/main/adapter/Group_Home_Main_Store_Kind_RecyclerviewAdapter.kt
@@ -38,7 +38,7 @@ class Group_Home_Main_Store_Kind_RecyclerviewAdapter(
             binding.ivStore.setAlpha(200)
             binding.tvStoreName.text = item.storeName
             binding.tvStoreLocation.text = item.address
-            binding.tvStoreRating.text = item.rating.toInt().toString()
+            binding.tvStoreRating.text = item.rating.toString()
             binding.toggleStoreLike.isChecked = item.isLiked == "Y"
 
             binding.clMainStoreRec.setOnClickListener {
